\name{base.error.paired}
\alias{base.error.paired}
\title{Local Pooled Error Test for Paired Data}
\description{
This is internal function in lpe.paired.
}
\usage{
base.error.paired(x, design,est.A, estimator, q, data.type)
}

\references{
Cho H, Smalley DM, Ross MM, Theodorescu D, Ley K and Lee JK (2007).  Statistical Identification of Differentially Labelled Peptides from Liquid Chromatography Tandem Mass Spectrometry, Proteomics, 7:3681-3692.
 }


\author{
HyungJun Cho and Jae K. Lee 
}

\seealso{
    \code{\link{lpe.paired.default}}
}

\keyword{internal}


