\name{plateletSet}
\alias{plateletSet}
\title{
LCMS proteomic data for platelte MPs 
}
\description{
 This data set consists of  LC-MS/MS data with three replicates of paired samples.
}

\source{
Garcia BA, Smalley DM, Cho H, Shabanowitz J, Ley K and Hunt DF (2005). The Platelet Microparticle Proteome, Journal of Proteome Research, 4:1516-1521.
}

\keyword{datasets}

