\name{print.lpe.paired.fdr}
\alias{print.lpe.paired.fdr}
\title{Local Pooled Error Test for Paired Data}
\description{
This prints lpe.paired.fdr
}
\usage{
\method{print}{lpe.paired.fdr}(x,...)
}

\arguments{
   \item{x}{an object created from  lpe.paired}
   \item{...}{other arguments}
}

\references{
Cho H, Smalley DM, Ross MM, Theodorescu D, Ley K and Lee JK (2007).  Statistical Identification of Differentially Labelled Peptides from Liquid Chromatography Tandem Mass Spectrometry, Proteomics, 7:3681-3692.
 }


\author{
HyungJun Cho and Jae K. Lee 
}



\keyword{internal}


