\name{summary.lpe.paired.fdr}
\alias{summary.lpe.paired.fdr}
\title{Local Pooled Error Test for Paired Data}
\description{
This summarize lpe.paired.
}
\usage{
\method{summary}{lpe.paired.fdr}(object,...)
}

\arguments{
   \item{x}{an object created from  lpe.paired}
   \item{...}{other arguments}
}


\references{
Cho H, Smalley DM, Ross MM, Theodorescu D, Ley K and Lee JK (2007).  Statistical Identification of Differentially Labelled Peptides from Liquid Chromatography Tandem Mass Spectrometry, Proteomics, 7:3681-3692.
 }


\author{
HyungJun Cho and Jae K. Lee 
}


\keyword{internal}


